Fix: weeklyControllerをlaundriesControllerのindexとして移動
